Added input validation file
What changes you have made?
.....

added isInputsValid function which will check the command line inputs
are valid or not
modified the main file
added isNumeric isEqual utility functions

Why this change is necessary?
....

to avoid unnecessary breakage of the code for invalid options

What is the impact of the change?
.....

if the user enter any invalid options the code will not break and it
will give a warning message