changed data flow and added testcases
What changes you have made?
.....
formatting message for save and query moved to processInputs function
added test cases for input validation

Why this change is necessary?
....
for making sure that code will not break for invalid inputs
if the result data flow is in object or array format further changes
will be easy if any

What is the impact of the change?
.....
code will be more secured from unnecessarry breakage
we can reuse the resultant data