Version:2 Changed the datastructure of database
What changes you have made?
.....
changed the datastructure of database from object of empId based array
of transactions to simple array of all the transactions

Why this change is necessary?
....
for query based on diffferent aspects it will be very useful

What is the impact of the change?
.....
dataStorage structure changed
we can apply different types of filter for different queries easily